refactor(sessions): use canonical lifecycle owners
